aio: ensure transport stores peername
When a socket is passed into loop.create_datagram_endpoint, the remote
address (peername) is always set to None which breaks the assumption of
DatagramClient that sendto can be called without an address.  Fix this
by forcing the transport to store the remote address.

This was a larger issue prior to
https://github.com/python/cpython/commit/63deaa5b70108ef441c57728322da6b4321db4fc
as it would prevent sending at all.  However, with that fix, the stored
address is largely ignored by by the transport aside from some error
checking and the peername will be used when the address is unspecified.