aio: raise TransportClosed from send/recv
When the transport backing a DatagramStream is closed, raise an
exception from recv() so that it doesn't block forever.  As a
convenience, do the same from send() to get a slightly nicer traceback.

Note that I did consider unifying the two queues into one, but that
would require the exception property and send() to be able to peek an
arbitrary distance into the queue which would lead to a much larger
refactor to use a different container.

Also note the comment in the test.  Handling multiple co-existing recv()
calls is not something we address in this commit.  If required, it can
easily be done by creating a dispatcher/adapter on top of the
DatagramStream.

closes #9