switch tooling to uv
PEP 517 is 7 years old now, and the tooling managing this repo was
already old then.  Now I'm old and tired and I don't want fight the lack
of a reasonable pyenv installer or manage my own virtualenvs or figure
out what the newest tool for uploading packages is.

uv gives me 2 out of 3 and as Meatloaf says, that ain't bad.

Fixes #20