make: remove last usage of setup.py
Given the switch to uv, there isn't much point in having setup.py around
anymore.  It's also not playing nicely in CI if I don't install the
system version of setuptools first.  The only point of this was so I
could copy Makefiles around and not have to change PACKAGE.